+++ src/MelloSilveiraTools.Mathematics/Models/VariableBoundaryValues.cs
@@ -1,0 +1,3 @@
+namespace SoftTissue.DataContracts.Commands.MechanicalModels.Commons;
+
+public sealed class VariableBoundaryValues : Dictionary<string, BoundaryValues>;